Switch entry/transaction naming in LocalLedger
