+++ apps/local_ledger/lib/local_ledger/errors/invalid_amount_error.ex
@@ -1,4 +1,4 @@
 defmodule LocalLedger.Errors.InvalidAmountError do
-  defexception message: "Entry could not be created.
+  defexception message: "Transaction could not be created.
                          Debit and Credit amounts were different."
 end
