+++ apps/local_ledger/lib/local_ledger/errors/same_address_error.ex
@@ -1,5 +1,5 @@
 defmodule LocalLedger.Errors.SameAddressError do
-  defexception message: "Entry could not be created. Some of the addresses are identical."
+  defexception message: "Transaction could not be created. Some of the addresses are identical."
 
   def error_message(addresses) do
     addresses = Enum.join(addresses, ", ")
